requirements.txt
azure-functions
pandas
azure-identity
azure-storage-blob
